Use ecs-deploy to deploy to ECS reliably
A task-definition in ECS using a Docker image tag like `latest` won't automatically
update when a new `latest` image is pushed, as I thought. Instead, the task definition
needs a new revision and the respective service needs to be updated to use the new
revision, which will then start a deployment.

ecs-deploy handles rolling releases out box, so we should have downtime-less deployments.